refactor: status label component (HL-908)
