+++ frontend/benefit/handler/src/components/statusLabel/StatusLabel.sc.ts
@@ -1,0 +1,14 @@
+import { APPLICATION_STATUSES } from 'benefit-shared/constants';
+import { Tag } from 'hds-react';
+import styled from 'styled-components';
+
+import { statusColorMap } from './utils';
+
+type StatusLabelProps = {
+  status?: APPLICATION_STATUSES;
+};
+
+export const $StatusLabel = styled(Tag)<StatusLabelProps>`
+  background-color: ${(props) =>
+    props.theme.colors[statusColorMap(props.status)]};
+`;
